examples/python/requirements.txt
redis>=5.0.0
redis-entraid>=1.0.0
python-dotenv>=1.0.0
